Introduce assault cards
Why:

We'd like to introduce a game mechanic that makes a distinction between
maneuvering a piece and assaulting another players piece. That is, we'd
like some tactic cards to only allow a piece to move, but not take an
opponent piece and some to only allow taking an opponent piece and some
to allow both moving and taking.

This commit:

Introduces a flew assault based tactic cards.

Updates maneuver cards to not allow maneuvering onto other pieces.

We additionally add state for resource and acquire cost of cards, which
we plan to use in a future commit.