Update cards to show more info
Why:

We'd like to improve the card styling to make it more clear what they
do.

This commit:

Updates the card component show the card description, resource
modifications, and kind (strategy or tactic). We update the styling to
to be more pleasing and pull out a common Card.tsx component to reuse in
the various game elements.