Introduce home row
Why:

We want have a concept of a home row as a part of the game mechanics.
Players deploy pieces to the home row, when players enter their
opponents home row they gain hegemony, etc.

This commit:

Adds a concept home row to game. We update some tests to allow us to
parameterize the size of the game board.