Update most packages to latest
Why:

We'd like to keep our dependencies up to date, as a matter of best
practice.

This commit:

Runs `ncu --upgrade` and takes all updates other than eslint, which we
leave at 8.49.0 as updating eslint to 9.1 (current release) would
require updating our eslint config and we'd like to tackle that in a
stand alone commit.