Introduce placement points
Why:

We'd like to allow players to add a piece every turn without the need to
play a card.

This commit:

Introduces a new turn point, the placement point, which controls how
many pieces a player can place. Players start every turn with 1
placement point. We update the DeployHoplite card to increase the
number of placement points the player has rather than immediately
requiring the player to select a cell to place a piece and we update the
selectCell game action logic to only look if the player has enough
placement points, separating logic for card playing logic.