Style turn points and sort supply cards
Why:

We'd like to improve the visual usability of the game. Currently the
supply pile cards don't all fit fully on the screen and the turn points
are hard to read.

This commit:

Updates the styling of the supply pile such that more cards can be seen
at once.
Updates the styling of the turn points so there is more contrast and
spacing around the point values. We introduce coloring to associate the
strategy points with the strategy cards and the tactic points with the
tactic cards.