Include e2e specs in validate spec
Why:

We want to include e2e specs in our validation script so that we are
more likely to run it more often and catch regressions local since we do
not have remote CI running for the project.

This commit:

Adds `npm run test:e2e` to our `bin/validate.sh` script.