Set cards when starting game
Why:

We'd like for the game to be easy to play test different starting hand
configurations and different supply card configurations. Additionally
we'd it to be easy to write end to end test for specific cards and by
allowing a way to set the starting hand in game this greatly simplifies
the setup of those e2e specs.

This commit:

Adds the ability to set the starting hand and available supply piles for
game.