Introduce draw mechanic
Why:

As a core game mechanic, we would like for players to be able to draw
cards from their deck during their turn.

This commit:

Introduces the logic for playing cards that allow for drawing more
cards. We follow the existing pattern of increasing the `turnPoints`
pool of points with a new `drawPoints` which will allow the user to
click on the draw pile to spend a draw point and draw a card.

To make testing cards easier we add the ability to set the hand size so
that e2e tests can be written where it's guaranteed to start with a hand
that includes a particular card. This allow us to play test games with
different starting hand sizes.

In addition to the draw functionality we include two refactors:

1. We update naming of turn points to a better match intent and be
   easier to grep for:
   - placementPoints -> hoplitePoints
   - draw -> drawPoints

2. We move all the calls to `setGame` to the GameState module expose
   higher level functions to be used in the view layer.