Refactor turn points data model
Why:

We are using several different data models to capture the notion of a
`Points Pool`, that is the collection of game points that a player can
use to do various game actions such as placing hoplites, playing cards,
and acquiring new cards. We'd like to consolidate these separate data
models into a single place to improve code maintainability.

Additionally all of the interactions with this collection of game points
is moniodal in nature, so we have some opportunity to introduce the
standard moniod abstraction to improve the composability of the code.

This commit:

Introduces a PointsPool moniod to be used for turn point and card value.