Move draw pile to main play mats row
Why:

We'd like for our draw pile to be easier to interact with since we now
have cards which allow the player to draw on their turn.

This commit:

Moves the draw and discard pile into the same row as the play mat and
the commit resources mat. All the piles are given the same visual
treatment.