Allow randomized supply piles in game setup
Why:

We'd like to let the players select a random set of supply piles in the
game setup so that games are easier to start if the player does not have
a strong preference on their starting cards.

This commit:

Adds the logic to randomize the game setup supply setup.