Refactor cards into tactic and strategy
Why:

We are running into a lot of places where we want to have the ability to
manage strategy and tactic cards independently. For example, when
setting up the game we'd like to ensure we have at least a few tactic
and few strategy cards.

Currently most of our data modeling treats strategy and tactic cards a
simply `cards` which makes it hard to treat them independently. We'd
like to better separate the tactic cards from the strategy cards in the
data model.

This commit:

Introduces types of `TacticCard` and `StrategyCard` to the data model.